[dev] - Updates detail page presentation.
